requirements.txt
annotated-types==0.7.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
dill==0.4.0
fonttools==4.58.2
ibm-cloud-sdk-core==3.24.2
ibm-platform-services==0.66.2
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.3
mpmath==1.3.0
numpy==2.2.6
packaging==25.0
pbr==6.1.1
pillow==11.2.1
psutil==7.0.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyJWT==2.10.1
pylatexenc==2.10
pyparsing==3.2.3
pyspnego==0.11.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
qiskit==2.0.2
qiskit-aer==0.17.1
qiskit-ibm-runtime==0.40.1
requests==2.32.4
requests_ntlm==1.3.0
rustworkx==0.16.0
scipy==1.15.3
six==1.17.0
stevedore==5.4.1
symengine==0.13.0
sympy==1.14.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
